pyproject.toml
[tool.poetry]
name = "Deiteo"
version = "1.1.0"
description = "Spark on K8S"
authors = ["Thelin90@Github"]

[tool.poetry.dependencies]
python = "3.9.1"
pyspark = "3.0.1"
pre-commit = "^2.9.3"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "6.1.1"
ipdb = "0.13.3"
pytest-cov = "2.10.1"
